Turn the libc and libm rules into 'order-only' prequisites
These prerequisites are GNU Make terminology.

This change forces the libc rule to be executed first so that the headers that
the libm rule needs are available. The original setup was using 'normal'
prerequisites which occasionally resulted in bizarre build breakage,
especially on multi-core build hosts. Seen often on my 8-way SMP build
host.

Note that this doesn't impede parallelisation of each rule indepent of
the other. It just serializes the rules themselves.

This fixes: https://gitlab.redox-os.org/redox-os/relibc/issues/127