Fix regression in ld.so elf size calculation
In patch 1182d1200603c9f33ec9968a1a00df5ebd3aa7d8, I mistakingly added
the size of the gap to the total size of the binary, which was not
accurate. As the size of the binary was calculate by subtracting the
upperbound from the lower bound, thus all gaps in the middle are taking
into account.