fix(stdio): `perror` should be allowed to take NULL prefixes
According to the spec:
[perror] writes a sequence of characters to the standard error stream
thus: first (if s is not a null pointer and the character pointed to by
s is not the null character), the string pointed to by s followed by a
colon (:) and a space; then the appropriate error message string
followed by a new-line character. [...]

In other words `s` is allowed to be NULL or empty.
Other libc implementations follow through with this as well.