Handle prepended whitespace for hashbang
Related to: #199

This patch fixes the shebang handler to eat whitespace prepended to the
interpreter.

For example, this should work:

```sh
\#! /bin/sh

echo "Hello from Redox"
```

Interestingly, the original shebang included a prepended space.

The shebang should also handle interpreter arguments as the issue
mentions. I intend to (attempt to!) fix that as well, but I decided to
split the work into two patches in case the latter issue is too
difficult for me.

Useful sources:
* https://homepages.cwi.nl/~aeb/std/hashexclam.html
* https://en.wikipedia.org/wiki/Shebang_%28Unix%29