Add strptime tests
relibc's behavior is similar to musl. I tested this suite for
relibc, musl, and glibc. The suite passes on musl and relibc but fails
on glibc because glibc sets tm_wday and tm_yday even if they're not
specified in the format string. I included, but commented out, the
values that would pass on glibc.