Implement BSD's err.h
err.h is a BSD extension to libc that simplifies displaying error
messages. The family of functions print the program binary name to a
file set by the user (or stderr if unset), an optional formatted string
like printf, and the error message associated with ERRNO or a user
provided value.

The functions are flexible but non-portable. glibc and musl both support
the family, but other libc implementations may not. glibc also
implements its own analogous, non-portable functions in error.h.