Ensure consistent allocator state on fork
This commit cleans up and reimplements old code intended to prevent a
deadlock on fork(). Allocating on fork() may deadlock if the allocator
is already locked. So, both the old code and this patch locks the
allocator on fork() then unlocks afterward for consistency.

The old code use pthread_atfork hooks. I manually lock and drop the
lock guard in fork() instead which is overall less code and easier to
follow. It's also helpful because we don't have to decide where to set
the hooks and don't have to rely on setting them on each spawned thread
(atfork hooks are thread local).