Add _LARGEFILE64_SOURCE (stat64 & friends)
_LARGEFILE64_SOURCE exposes 64-bit capable versions of functions like
stat or readdir. Redox, Linux, the BSDs, and other Unixes are
(obviously) 64-bit capable and have been that way for a long time.
_LARGEFILE64_SOURCE exists for backwards compatibility with C99 code.
Code targeting a standard beyond C99 is compiled with implicit
_LARGEFILE64_SOURCE support.

Musl, glibc, and other implementations simply define the 64-bit
functions as the original function - e.g. stat64 IS stat. Special
versions of the functions aren't needed anymore, and the defines are
provided by backward compatibility.

The tl;dr is that this just adds #define func64 func gated by
_LARGEFILE64_SOURCE to a few headers.